Remove util function from package
